+++ Source/swiftlint/Helpers/LintOrAnalyzeCommand.swift
@@ -28,7 +28,12 @@ enum LintOrAnalyzeMode {
 struct LintOrAnalyzeCommand {
     static func run(_ options: LintOrAnalyzeOptions) async throws {
         if options.inProcessSourcekit {
-            SourceKittenConfiguration.preferInProcessSourceKit = true
+            queuedPrintError(
+                """
+                warning: The --in-process-sourcekit option is deprecated. \
+                SwiftLint now always uses an in-process SourceKit.
+                """
+            )
         }
         try await Signposts.record(name: "LintOrAnalyzeCommand.run") {
             try await options.autocorrect ? autocorrect(options) : lintOrAnalyze(options)